+++ google-cloud-bigtable/src/main/java/com/google/cloud/bigtable/data/v2/stub/metrics/MetricsTracer.java
@@ -129,8 +129,8 @@ class MetricsTracer extends BigtableTracer {
   }
 
   @Override
-  public void attemptStarted(int i) {
-    attempt = i;
+  public void attemptStarted(int attemptNumber) {
+    attempt = attemptNumber;
     attemptCount++;
     attemptTimer = Stopwatch.createStarted();
     attemptResponseCount = 0;
@@ -211,7 +211,7 @@ class MetricsTracer extends BigtableTracer {
   }
 
   @Override
-  public void recordGfeMetadata(@Nullable Long latency) {
+  public void recordGfeMetadata(@Nullable Long latency, @Nullable Throwable throwable) {
     MeasureMap measures = stats.newMeasureMap();
     if (latency != null) {
       measures
@@ -220,6 +220,18 @@ class MetricsTracer extends BigtableTracer {
     } else {
       measures.put(RpcMeasureConstants.BIGTABLE_GFE_HEADER_MISSING_COUNT, 1L);
     }
+    measures.record(
+        newTagCtxBuilder()
+            .putLocal(RpcMeasureConstants.BIGTABLE_STATUS, Util.extractStatus(throwable))
+            .build());
+  }
+
+  @Override
+  public void batchRequestThrottled(long totalThrottledMs) {
+    MeasureMap measures =
+        stats
+            .newMeasureMap()
+            .put(RpcMeasureConstants.BIGTABLE_BATCH_THROTTLED_TIME, totalThrottledMs);
     measures.record(newTagCtxBuilder().build());
   }